requirements/_base.txt
django<1.5
bpython==0.12
django-braces==0.2.1
django-model-utils==1.1.0
South==0.7.6
psycopg2==2.4.6
django-gravatar2==1.0.6
django-crispy-forms==1.2.3
